+++ ft_int_length.c
@@ -6,13 +6,13 @@
 /*   By: rschlott <rschlott@student.42wolfsburg.de> +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/05/22 18:56:04 by rschlott          #+#    #+#             */
-/*   Updated: 2022/05/22 19:04:52 by rschlott         ###   ########.fr       */
+/*   Updated: 2022/05/23 14:55:14 by rschlott         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "ft_printf.h"
 
-int ft_int_length(int n)
+int ft_int_length(long long n)   // long int: so that neg unsigned int length gets counted correctly!
 {
     int length;
